update folder names
