+++ jsconfig.json
@@ -5,7 +5,7 @@
 			"@/components/*": ["components/*"],
 			"@/config/*": ["config/*"],
 			"@/utils/*": ["utils/*"],
-			"@/lib/*": ["lib/*"]
+			"@/lib/*": ["site_data/*"]
 		}
 	}
 }
